+++ src/duck_ingestor.py
